requirements.txt
pandas==2.2.3
numpy==2.1.3
statsmodels==0.14.4
scikit-learn==1.5.2
psycopg2-binary==2.9.10
yfinance==0.2.54
polygon-api-client==1.14.2
python-dotenv==1.0.1
pytest==8.3.4
pyyaml==6.0.2
pyarrow==18.1.0
matplotlib==3.10.0
pandas-ta==0.3.14b0
imbalanced-learn==0.12.4
joblib==1.4.2
lightgbm==4.5.0
shap==0.46.0
